@@ -8,103 +8,156 @@ abstract class FlutterFlowTheme {
     return LightModeTheme();
   }
 
-  Color primaryColor;
-  Color secondaryColor;
-  Color tertiaryColor;
-  Color alternate;
-  Color primaryBackground;
-  Color secondaryBackground;
-  Color primaryText;
-  Color secondaryText;
+  @Deprecated('Use primary instead')
+  Color get primaryColor => primary;
+  @Deprecated('Use secondary instead')
+  Color get secondaryColor => secondary;
+  @Deprecated('Use tertiary instead')
+  Color get tertiaryColor => tertiary;
 
-  Color grayDark;
-  Color dark900;
-  Color background;
-  Color grayIcon;
-  Color tertiary;
-  Color gray200;
-  Color gray600;
-  Color black600;
-  Color tertiary400;
-  Color textColor;
-  Color maximumBlueGreen;
-  Color plumpPurple;
-  Color platinum;
-  Color ashGray;
-  Color darkSeaGreen;
-  Color primaryBtnText;
-  Color lineColor;
-  Color btnText;
-  Color customColor3;
-  Color customColor4;
-  Color white;
+  late Color primary;
+  late Color secondary;
+  late Color tertiary;
+  late Color alternate;
+  late Color primaryText;
+  late Color secondaryText;
+  late Color primaryBackground;
+  late Color secondaryBackground;
+  late Color accent1;
+  late Color accent2;
+  late Color accent3;
+  late Color accent4;
+  late Color success;
+  late Color warning;
+  late Color error;
+  late Color info;
 
-  String get title1Family => typography.title1Family;
-  TextStyle get title1 => typography.title1;
-  String get title2Family => typography.title2Family;
-  TextStyle get title2 => typography.title2;
-  String get title3Family => typography.title3Family;
-  TextStyle get title3 => typography.title3;
-  String get subtitle1Family => typography.subtitle1Family;
-  TextStyle get subtitle1 => typography.subtitle1;
-  String get subtitle2Family => typography.subtitle2Family;
-  TextStyle get subtitle2 => typography.subtitle2;
-  String get bodyText1Family => typography.bodyText1Family;
-  TextStyle get bodyText1 => typography.bodyText1;
-  String get bodyText2Family => typography.bodyText2Family;
-  TextStyle get bodyText2 => typography.bodyText2;
+  late Color primaryBtnText;
+  late Color lineColor;
+
+  @Deprecated('Use displaySmallFamily instead')
+  String get title1Family => displaySmallFamily;
+  @Deprecated('Use displaySmall instead')
+  TextStyle get title1 => typography.displaySmall;
+  @Deprecated('Use headlineMediumFamily instead')
+  String get title2Family => typography.headlineMediumFamily;
+  @Deprecated('Use headlineMedium instead')
+  TextStyle get title2 => typography.headlineMedium;
+  @Deprecated('Use headlineSmallFamily instead')
+  String get title3Family => typography.headlineSmallFamily;
+  @Deprecated('Use headlineSmall instead')
+  TextStyle get title3 => typography.headlineSmall;
+  @Deprecated('Use titleMediumFamily instead')
+  String get subtitle1Family => typography.titleMediumFamily;
+  @Deprecated('Use titleMedium instead')
+  TextStyle get subtitle1 => typography.titleMedium;
+  @Deprecated('Use titleSmallFamily instead')
+  String get subtitle2Family => typography.titleSmallFamily;
+  @Deprecated('Use titleSmall instead')
+  TextStyle get subtitle2 => typography.titleSmall;
+  @Deprecated('Use bodyMediumFamily instead')
+  String get bodyText1Family => typography.bodyMediumFamily;
+  @Deprecated('Use bodyMedium instead')
+  TextStyle get bodyText1 => typography.bodyMedium;
+  @Deprecated('Use bodySmallFamily instead')
+  String get bodyText2Family => typography.bodySmallFamily;
+  @Deprecated('Use bodySmall instead')
+  TextStyle get bodyText2 => typography.bodySmall;
+
+  String get displayLargeFamily => typography.displayLargeFamily;
+  TextStyle get displayLarge => typography.displayLarge;
+  String get displayMediumFamily => typography.displayMediumFamily;
+  TextStyle get displayMedium => typography.displayMedium;
+  String get displaySmallFamily => typography.displaySmallFamily;
+  TextStyle get displaySmall => typography.displaySmall;
+  String get headlineLargeFamily => typography.headlineLargeFamily;
+  TextStyle get headlineLarge => typography.headlineLarge;
+  String get headlineMediumFamily => typography.headlineMediumFamily;
+  TextStyle get headlineMedium => typography.headlineMedium;
+  String get headlineSmallFamily => typography.headlineSmallFamily;
+  TextStyle get headlineSmall => typography.headlineSmall;
+  String get titleLargeFamily => typography.titleLargeFamily;
+  TextStyle get titleLarge => typography.titleLarge;
+  String get titleMediumFamily => typography.titleMediumFamily;
+  TextStyle get titleMedium => typography.titleMedium;
+  String get titleSmallFamily => typography.titleSmallFamily;
+  TextStyle get titleSmall => typography.titleSmall;
+  String get labelLargeFamily => typography.labelLargeFamily;
+  TextStyle get labelLarge => typography.labelLarge;
+  String get labelMediumFamily => typography.labelMediumFamily;
+  TextStyle get labelMedium => typography.labelMedium;
+  String get labelSmallFamily => typography.labelSmallFamily;
+  TextStyle get labelSmall => typography.labelSmall;
+  String get bodyLargeFamily => typography.bodyLargeFamily;
+  TextStyle get bodyLarge => typography.bodyLarge;
+  String get bodyMediumFamily => typography.bodyMediumFamily;
+  TextStyle get bodyMedium => typography.bodyMedium;
+  String get bodySmallFamily => typography.bodySmallFamily;
+  TextStyle get bodySmall => typography.bodySmall;
 
   Typography get typography => ThemeTypography(this);
 }
 
 class LightModeTheme extends FlutterFlowTheme {
-  Color primaryColor = const Color(0xFFF34A36);
-  Color secondaryColor = const Color(0xFFEE8B60);
-  Color tertiaryColor = const Color(0xFFFFFFFF);
-  Color alternate = const Color(0xFF39D2C0);
-  Color primaryBackground = const Color(0xFFF1F4F8);
-  Color secondaryBackground = const Color(0xFFFFFFFF);
-  Color primaryText = const Color(0xFF090F13);
-  Color secondaryText = const Color(0xFF57636C);
+  @Deprecated('Use primary instead')
+  Color get primaryColor => primary;
+  @Deprecated('Use secondary instead')
+  Color get secondaryColor => secondary;
+  @Deprecated('Use tertiary instead')
+  Color get tertiaryColor => tertiary;
 
-  Color grayDark = Color(0xFF616E78);
-  Color dark900 = Color(0xFF090F13);
-  Color background = Color(0xFF14181B);
-  Color grayIcon = Color(0xFF95A1AC);
-  Color tertiary = Color(0xFF39D2C0);
-  Color gray200 = Color(0xFFDBE2E7);
-  Color gray600 = Color(0xFF262D34);
-  Color black600 = Color(0xFF090F13);
-  Color tertiary400 = Color(0xFF39D2C0);
-  Color textColor = Color(0xFF1E2429);
-  Color maximumBlueGreen = Color(0xFF59C3C3);
-  Color plumpPurple = Color(0xFF52489C);
-  Color platinum = Color(0xFFEBEBEB);
-  Color ashGray = Color(0xFFCAD2C5);
-  Color darkSeaGreen = Color(0xFF84A98C);
-  Color primaryBtnText = Color(0xFFFFFFFF);
-  Color lineColor = Color(0xFFE0E3E7);
-  Color btnText = Color(0xFFFFFFFF);
-  Color customColor3 = Color(0xFFDF3F3F);
-  Color customColor4 = Color(0xFF090F13);
-  Color white = Color(0xFFFFFFFF);
+  late Color primary = const Color(0xFF4B39EF);
+  late Color secondary = const Color(0xFF39D2C0);
+  late Color tertiary = const Color(0xFFEE8B60);
+  late Color alternate = const Color(0xFFFA6650);
+  late Color primaryText = const Color(0xFF101213);
+  late Color secondaryText = const Color(0xFF2F80ED);
+  late Color primaryBackground = const Color(0xFFF1F4F8);
+  late Color secondaryBackground = const Color(0xFFFFFFFF);
+  late Color accent1 = const Color(0xFF616161);
+  late Color accent2 = const Color(0xFF757575);
+  late Color accent3 = const Color(0xFFE0E0E0);
+  late Color accent4 = const Color(0xFFEEEEEE);
+  late Color success = const Color(0xFF04A24C);
+  late Color warning = const Color(0xFFFCDC0C);
+  late Color error = const Color(0xFFE21C3D);
+  late Color info = const Color(0xFF1C4494);
+
+  late Color primaryBtnText = Color(0xFFFFFFFF);
+  late Color lineColor = Color(0xFFE0E3E7);
 }
 
 abstract class Typography {
-  String get title1Family;
-  TextStyle get title1;
-  String get title2Family;
-  TextStyle get title2;
-  String get title3Family;
-  TextStyle get title3;
-  String get subtitle1Family;
-  TextStyle get subtitle1;
-  String get subtitle2Family;
-  TextStyle get subtitle2;
-  String get bodyText1Family;
-  TextStyle get bodyText1;
-  String get bodyText2Family;
-  TextStyle get bodyText2;
+  String get displayLargeFamily;
+  TextStyle get displayLarge;
+  String get displayMediumFamily;
+  TextStyle get displayMedium;
+  String get displaySmallFamily;
+  TextStyle get displaySmall;
+  String get headlineLargeFamily;
+  TextStyle get headlineLarge;
+  String get headlineMediumFamily;
+  TextStyle get headlineMedium;
+  String get headlineSmallFamily;
+  TextStyle get headlineSmall;
+  String get titleLargeFamily;
+  TextStyle get titleLarge;
+  String get titleMediumFamily;
+  TextStyle get titleMedium;
+  String get titleSmallFamily;
+  TextStyle get titleSmall;
+  String get labelLargeFamily;
+  TextStyle get labelLarge;
+  String get labelMediumFamily;
+  TextStyle get labelMedium;
+  String get labelSmallFamily;
+  TextStyle get labelSmall;
+  String get bodyLargeFamily;
+  TextStyle get bodyLarge;
+  String get bodyMediumFamily;
+  TextStyle get bodyMedium;
+  String get bodySmallFamily;
+  TextStyle get bodySmall;
 }
 
 class ThemeTypography extends Typography {
@@ -112,72 +165,128 @@ class ThemeTypography extends Typography {
 
   final FlutterFlowTheme theme;
 
-  String get title1Family => 'Lexend Deca';
-  TextStyle get title1 => GoogleFonts.getFont(
-        'Lexend Deca',
-        color: theme.tertiaryColor,
+  String get displayLargeFamily => 'Montserrat';
+  TextStyle get displayLarge => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.normal,
+        fontSize: 57.0,
+      );
+  String get displayMediumFamily => 'Montserrat';
+  TextStyle get displayMedium => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.normal,
+        fontSize: 45.0,
+      );
+  String get displaySmallFamily => 'Montserrat';
+  TextStyle get displaySmall => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
         fontWeight: FontWeight.w600,
-        fontSize: 24,
+        fontSize: 24.0,
       );
-  String get title2Family => 'Lexend Deca';
-  TextStyle get title2 => GoogleFonts.getFont(
-        'Lexend Deca',
-        color: theme.tertiaryColor,
+  String get headlineLargeFamily => 'Montserrat';
+  TextStyle get headlineLarge => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.normal,
+        fontSize: 32.0,
+      );
+  String get headlineMediumFamily => 'Montserrat';
+  TextStyle get headlineMedium => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.secondaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 22.0,
+      );
+  String get headlineSmallFamily => 'Montserrat';
+  TextStyle get headlineSmall => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 20.0,
+      );
+  String get titleLargeFamily => 'Montserrat';
+  TextStyle get titleLarge => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
         fontWeight: FontWeight.w500,
-        fontSize: 22,
+        fontSize: 22.0,
       );
-  String get title3Family => 'Lexend Deca';
-  TextStyle get title3 => GoogleFonts.getFont(
-        'Lexend Deca',
-        color: theme.tertiary,
-        fontWeight: FontWeight.bold,
-        fontSize: 20,
+  String get titleMediumFamily => 'Montserrat';
+  TextStyle get titleMedium => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 18.0,
       );
-  String get subtitle1Family => 'Lexend Deca';
-  TextStyle get subtitle1 => GoogleFonts.getFont(
-        'Lexend Deca',
-        color: theme.dark900,
+  String get titleSmallFamily => 'Montserrat';
+  TextStyle get titleSmall => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.secondaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 16.0,
+      );
+  String get labelLargeFamily => 'Montserrat';
+  TextStyle get labelLarge => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
         fontWeight: FontWeight.w500,
-        fontSize: 18,
+        fontSize: 14.0,
       );
-  String get subtitle2Family => 'Lexend Deca';
-  TextStyle get subtitle2 => GoogleFonts.getFont(
-        'Lexend Deca',
-        color: theme.primaryBackground,
-        fontWeight: FontWeight.normal,
-        fontSize: 16,
+  String get labelMediumFamily => 'Montserrat';
+  TextStyle get labelMedium => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w500,
+        fontSize: 12.0,
       );
-  String get bodyText1Family => 'Lexend Deca';
-  TextStyle get bodyText1 => GoogleFonts.getFont(
-        'Lexend Deca',
-        color: theme.grayIcon,
-        fontWeight: FontWeight.normal,
-        fontSize: 14,
+  String get labelSmallFamily => 'Montserrat';
+  TextStyle get labelSmall => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w500,
+        fontSize: 11.0,
       );
-  String get bodyText2Family => 'Lexend Deca';
-  TextStyle get bodyText2 => GoogleFonts.getFont(
-        'Lexend Deca',
-        color: theme.grayIcon,
+  String get bodyLargeFamily => 'Montserrat';
+  TextStyle get bodyLarge => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
         fontWeight: FontWeight.normal,
-        fontSize: 12,
+        fontSize: 16.0,
+      );
+  String get bodyMediumFamily => 'Montserrat';
+  TextStyle get bodyMedium => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 14.0,
+      );
+  String get bodySmallFamily => 'Montserrat';
+  TextStyle get bodySmall => GoogleFonts.getFont(
+        'Montserrat',
+        color: theme.secondaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 14.0,
       );
 }
 
 extension TextStyleHelper on TextStyle {
   TextStyle override({
-    String fontFamily,
-    Color color,
-    double fontSize,
-    FontWeight fontWeight,
-    double letterSpacing,
-    FontStyle fontStyle,
+    String? fontFamily,
+    Color? color,
+    double? fontSize,
+    FontWeight? fontWeight,
+    double? letterSpacing,
+    FontStyle? fontStyle,
     bool useGoogleFonts = true,
-    TextDecoration decoration,
-    double lineHeight,
+    TextDecoration? decoration,
+    double? lineHeight,
   }) =>
       useGoogleFonts
           ? GoogleFonts.getFont(
-              fontFamily,
+              fontFamily!,
               color: color ?? this.color,
               fontSize: fontSize ?? this.fontSize,
               letterSpacing: letterSpacing ?? this.letterSpacing,
