@@ -37,10 +37,10 @@ class FFPlace {
   bool operator ==(other) =>
       other is FFPlace &&
       latLng == other.latLng &&
-      name == name &&
-      address == address &&
-      city == city &&
-      state == state &&
-      country == country &&
-      zipCode == zipCode;
+      name == other.name &&
+      address == other.address &&
+      city == other.city &&
+      state == other.state &&
+      country == other.country &&
+      zipCode == other.zipCode;
 }
