@@ -1,16 +1,19 @@
 // Export pages
-export 'login/login_widget.dart' show LoginWidget;
-export 'register/register_widget.dart' show RegisterWidget;
-export 'clients/clients_widget.dart' show ClientsWidget;
-export 'complete_profile/complete_profile_widget.dart'
-    show CompleteProfileWidget;
-export 'forgot_password/forgot_password_widget.dart' show ForgotPasswordWidget;
-export 'visitor/visitor_widget.dart' show VisitorWidget;
-export 'edit_profile/edit_profile_widget.dart' show EditProfileWidget;
-export 'change_password/change_password_widget.dart' show ChangePasswordWidget;
-export 'tracks/tracks_widget.dart' show TracksWidget;
-export 'tracks_my/tracks_my_widget.dart' show TracksMyWidget;
-export 'profile_l_k/profile_l_k_widget.dart' show ProfileLKWidget;
-export 'track_record/track_record_widget.dart' show TrackRecordWidget;
-export 'new_online_track/new_online_track_widget.dart'
-    show NewOnlineTrackWidget;
+export '/pages/main_flow/login/login_widget.dart' show LoginWidget;
+export '/pages/main_flow/orders_list/orders_list_widget.dart'
+    show OrdersListWidget;
+export '/pages/main_flow/order/order_widget.dart' show OrderWidget;
+export '/pages/main_flow/photos/photos_widget.dart' show PhotosWidget;
+export '/pages/main_flow/status_move_meeting/status_move_meeting_widget.dart'
+    show StatusMoveMeetingWidget;
+export '/pages/main_flow/status/status_widget.dart' show StatusWidget;
+export '/pages/main_flow/photo_document/photo_document_widget.dart'
+    show PhotoDocumentWidget;
+export '/pages/main_flow/signing_sms/signing_sms_widget.dart'
+    show SigningSmsWidget;
+export '/pages/main_flow/signing_client/signing_client_widget.dart'
+    show SigningClientWidget;
+export '/pages/main_flow/signing_agent/signing_agent_widget.dart'
+    show SigningAgentWidget;
+export '/pages/main_flow/binding/binding_widget.dart' show BindingWidget;
+export '/pages/main_flow/dop_main/dop_main_widget.dart' show DopMainWidget;
